Implemented video detail page display
This implements the video detail page display, using a few material
design components for some of the UI elements. This also allows the user
to edit the title and description for the video on this page, shows
error and status messages if the transcoding failed or is in progress,
etc.

closes #135
closes #136