Updated settings and requirements to fix deployment issues
The file parsing available in uWSGI has a hardcoded line limit which was causing problems with managing the certificate strings via the env file. Added a block to read in the env file and set environment variables automatically as part of settings.py to work around that issue.

Removed the `-e` flags from the git dependencies so that they get installed to the `dist-packages` directory instead of the home directory of the user that runs the pip command. This was causing issues with being able to import those packages at runtime.

Added the `no-binary` flag for psycopg2 due to issues with how the wheel is linked to OpenSSL for Debian 9 (https://github.com/unbit/uwsgi/issues/1569)

Added the option to set the celery broker separately from the results backend in order to allow for using RabbitMQ in place of Redis.